requirements.txt
keras-nightly
pydub
carla
openai
opencv-python
numpy
SpeechRecognition
pygame
tensorflow
ffmpeg-downloader
google-generativeai